Only delete furlfile if we created it
git-svn-id: http://usg.lofar.org/svn/code/trunk@4949 e3a606d5-b720-0410-b6ad-cb58f335abf8

+++ src/pipeline/deploy/ipengine.sh
@@ -11,6 +11,7 @@ case "$2" in
   start) 
          if [ $FURL ]; then
              FURLFILE=`mktemp`
+             TMPFURL=1
              echo $FURL > $FURLFILE
          else
              FURLFILE=$CONTROLPATH/engine.furl
@@ -43,7 +44,9 @@ case "$2" in
                     echo "ipengine already running on `hostname`"
                 fi
             done
-         rm $FURLFILE
+         if [ $TMPFURL ]; then
+             rm $FURLFILE
+         fi
          ;;
   stop)
          for PIDFILE in $PIDPATH/ipengine*.pid
